Correct the exponent value calculated when converting from float to half
When there is overflow, the exponent is set to all 1s (0x1f).
This is equal to 31 in decimal. The original code set the exponent
to hex (0x31).

The test for checking the exponent when there is an overflow has
been added to android.util.cts.HalfTest in another patch.

Test: cts-tradefed  run singleCommand cts-dev --module CtsUtilTestCases --test android.util.cts.HalfTest

Change-Id: I945b9aca976b2c85bce371ed9002a8c6bc40f4ff
Signed-off-by: DennySPB <dennyspb@gmail.com>